added sample widget and export functions
