+++ client/src/components/Widgets/TextToSpeechWidget/TTSTools.js
@@ -1,0 +1,23 @@
+const readFileSync = require("fs").readFileSync;
+const sdk = require("microsoft-cognitiveservices-speech-sdk");
+
+const speechConfig = sdk.SpeechConfig.fromSubscription(
+  "6fe10de8d0ca4bd894e02d8c2a2debb2",
+  "westus"
+);
+const audioConfig = sdk.AudioConfig.fromSpeakerOutput();
+const synthesizer = new sdk.SpeechSynthesizer(speechConfig, audioConfig);
+
+export function microsoft(text) {
+  synthesizer.speakTextAsync(
+    text,
+    (result) => {
+      if (result) {
+      }
+    },
+    (error) => {
+      console.log(error);
+      synthesizer.close();
+    }
+  );
+}
